requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
openai
google-cloud-aiplatform
loguru
prometheus-client
pdfplumber
python-multipart
pytesseract
pydantic-settings
google-genai
httpx
pytest
pytest-asyncio
pytest-cov
pytest-mock
mongomock
aiofiles    